@@ -1,10 +1,10 @@
 name: css_text
 description: This library allows you to use simple CSS inline styles to style your text.
-version: 2.1.0
-homepage: https://github.com/hathibelagal-dev/css_text_for_flutter
+version: 3.0.0
+homepage: https://github.com/slarrauri/css_text_for_flutter
 
 environment:
-  sdk: ">=2.10.3 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
