@@ -27,7 +27,7 @@ class HTML {
   /// click events on anchor tags. The function will receive the actual link
   /// as its argument.
   static TextSpan toTextSpan(BuildContext context, String htmlContent,
-      {Function linksCallback}) {
+      {Function? linksCallback}) {
     Parser p = Parser(context, htmlContent, linksCallback: linksCallback);
     return TextSpan(text: "", children: p.parse());
   }
@@ -38,7 +38,7 @@ class HTML {
   /// click events on anchor tags. The function will receive the actual link
   /// as its argument.
   static RichText toRichText(BuildContext context, String htmlContent,
-      {Function linksCallback}) {
+      {Function? linksCallback}) {
     return RichText(
         text: toTextSpan(context, htmlContent, linksCallback: linksCallback));
   }
